fix: modal width fix for left pane open
